# Improve performance with large number of queued prompts