Mitigates https://github.com/Comfy-Org/ComfyUI_frontend/issues/2435

ComfyUI's web app exhibits significant lag when a large number of prompts are queued. I verified what was found in the linked thread, pinpointing the slowdown to the specific call `api.getQueue()` as the number of prompts grows. Every time a prompt is submitted, the front-end synchronizes the entire queue state by making this call.

Profiling the back-end, I found most of the time is spent executing `copy.deepcopy` inside PromptQueue while a mutex is held.

![GetQueue_Before](https://github.com/user-attachments/assets/720b59c1-c836-469a-ba99-3ec6b8fbbe5c)

This PR adds a new version of PromptQueue's `get_current_queue` method that performs `copy.copy` instead of `copy.deepcopy`. Allowing read-only shared access is safe as long as `PromptQueue` treats its queue items as immutable, which it does by returning deep copies in its other methods. I named the new version `get_current_queue_volatile` to remind callers they must only read and release these values.

Using the updated method allows Comfy to respond much faster to requests to the `/queue` route. The web app remains usable on my machine with several hundred queue items. I enqueued 300 prompts for these benchmarks.

Notice how the whole `deepcopy` block is gone, and you can actually see Comfy working on other stuff:
![After change](https://github.com/user-attachments/assets/d636cab8-1fb4-49b3-a8da-aa2105ed30d7)


Caching the JSON representations may be another 80/20 fix. I could look into that in the future. A more complex but optimized solution would be to rewrite state sync so the front end only requests a diff with updates to the queue instead of a full sync. Still, this quick fix helps a lot.