@catboxanon 2025-05-18
Nice work looking into this with benchmarks to show.

I had [something similar](https://github.com/comfyanonymous/ComfyUI/pull/6772) opened months ago with zero activity but I'll go ahead and close that since I think your implementation makes more sense.