@christian-byrne 2025-05-20
This is great. I just did basic test:

1. Copy and paste the default workflow on top of itself to make large graph
2. Queue 100 times

With this PR:

1. Response times on the `/queue` and `/history` endpoints stays under 100ms
2. Takes ~3-5s for all jobs to be queued fully
3. Performance is slightly laggy

Without this PR:

1. Response times on the `/queue` and `/history` endpoints gradually increase from 100ms to 500ms, eventually going over 1000ms
2. Gradually decreases speed as jobs are queued. Is not even able to queue all 100 jobs eargerly, as the response times and lag become too much
3. Performance on client is slightly worse, not super noticeable difference on my machine though. Did not compare inference speed, only UI responsiveness

I can do some other tests later.