@christian-byrne 2025-05-20
Tested queue features:

- Deleting individual items from history 
- Cancelling individual items items in queue
- Cancelling all items in queue
- Queueing large batch at once
- Cancelling large batch then re-queueing batch


https://github.com/user-attachments/assets/08232830-de91-4a4c-a4c4-8264c7e6562d