@0X-JonMichaelGalindo 2025-05-21
> I queue ~400 gens before work (assets, game elements, & prototypes, doing best of 100 generation). This is epic. Thank you!