@mikaizhao-hue 2025-08-01
The same issue still exists when using the comfyui desktop. Performance goes down rapidly when queued more than 100