@christian-byrne 2025-08-02
Which desktop version? The performance still scales with queue size just less so. There may also be other bottlenecks besides copying queue items.