@mikaizhao-hue 2025-08-05
> Which desktop version? The performance still scales with queue size just less so. There may also be other bottlenecks besides copying queue items.

I am using nightly version Comfyui desktop updated on 4th August and ComfyUI v0.3.48, when queue tasks over 100, it starts lagging and almost freezes when it goes to 200-300