@christian-byrne 2025-08-07
I'm not aware of any changes from that period that could cause that. It is also influenced by the number of items in your input folder and the size of your workflows.