@kicapanmanis07-ux 2025-08-19
I just updated from ComfyUI 0.3.10-40 / Frontend 1.6.18 to ComfyUI 0.3.50 / Frontend 1.23.4 and now queing up is really really really slow when I try to que thousands in advance. It starts to slow down after 300 is in the queue.

In my old version it will also be slow but not if I just opened the server and the browser, where it can que thousands really fast. Now it is extremely slow. Its a shame because aside from this issue, the new version is better in every way but sadly it is very difficult to use when queing up before I leave the PC now.