@gonztok 2025-08-20
> > It actually started happening after one update in mid of May. I used to queue 200-300 tasks and leave them running overnight without issue, but after that update, it started lagging and then eventually froze after queuing more than 100 items.
> 
> @mikaizhao-hue See if this extenstion will work for you. It has different approach to internal queue storage and from my tests so far can take large amount of queue items without hiccups - among other features. https://github.com/QuietNoise/ComfyUI-Queue-Manager

Thanks! This is working wonders for me!